NEW Game - Sequence Distractor Match to Sample (+ NI Script)
Sequence Distractor Match to Sample (SDMS)
On screen the user is presented a slot machine UI with 3 windows in the center of the screen, and 4 buttons below.

The task is split into 3 phases.
- Cue
- Delay
- Recall

During the Cue phase, 3 stimuli are shown to the user, one in each window.

During the delay phase there is a 50% chance for the stimuli to "shuffle" showing multiple distractors throughout the delay duration.

During the recall phase, the user must indicate how many stimuli are the correct size/shape/pattern AND in the correct location.

To select their answer the user will use the 4 buttons under the 3 windows, centered in the lower middle of the screen.

X: No stimulus is in the same position with the same visual features

1 (single diamond): Only one stimulus is in the same position and has the same features

2 (dual cherries): Exactly 2 stimuli are in the same position with the same features

3 (trio banana bunch): All 3 stimuli are in the same position with the same features.

The trial set was defined in this way:
- Balance all factors (Distractor, MatchLevel, Serial Slot) so each stimulus appears once per cell, using a Williams Latin square to generate 24 triads that evenly distribute stimuli across positions.

- This results in 192 trials where each condition combination was equally represented, avoiding repeats and run‑order biases.

Additionally, I have created another "ExtraFunctions" type script specific for our lab needs. This script uses our NativeDAQmx wrapper to create a task for the NI DAQ which will send TTL pulses to our hardware on events "StartEndBlock" and "TrialOn".

I have also created a prefab to easily add all functionality necessary for our lab needs. Just drag and drop the prefab (located in Assets/Scripts/NI DAQ/NIDAQ_CC_Complete)

This instantly makes any game (which is correctly configured with "clickable" objects having a stimulus ID) work with our eye tracking and recording setup.