Template instructions + "Correct" Logic Update
Each new task will have a template panel pop up when the "HOW TO PLAY" button is pressed.

Now, ExperimentDefinitions have an "IsSample" toggle that pops us if "IsStimulus" is true. This allows for the indicies to be cached to a "sampleIdxs" array used for deciding the "bool correct" choice during the task.

Also updated the Trial Definition Generator to autopopulate the specfic "Sample" options depending on the state's "IsSample" value.